+++ frontend/src/components/Accordion/Accordion.stories.tsx
@@ -1,9 +1,9 @@
 import { Meta, StoryObj } from '@storybook/react-webpack5';
-import Accodion from '.';
+import Accordion from '.';
 
-const meta: Meta<typeof Accodion> = {
-  title: 'Components/Accodion',
-  component: Accodion,
+const meta: Meta<typeof Accordion> = {
+  title: 'Components/Accordion',
+  component: Accordion,
   tags: ['autodocs'],
 
   decorators: [
@@ -17,7 +17,7 @@ const meta: Meta<typeof Accodion> = {
 
 export default meta;
 
-type Story = StoryObj<typeof Accodion>;
+type Story = StoryObj<typeof Accordion>;
 
 export const OpenAccodion: Story = {
   args: {
