+++ frontend/src/components/Accordion/Accordion.styled.ts
